removed single createSSMParameter calls and joined them to the bundle SSM create call fixed ssm parameters moved dashboard to lib hierarchy
